Added another XSD parser implementation to simplify development.
