+++ tsconfig.json
@@ -116,5 +116,5 @@
   },
 
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "**/*.test.ts"]
+  "exclude": ["node_modules", "dist"]
 }
